Sitemaker sites
This is files are a source of Sitemaker web-program which can help users in creating web-sites. User doesn't need to know difficult codes of website making. It's based on HTML5, CSS3,Javascript and jQuery.  It uses famous jQuery modules. Except some important modules are created by the author of this project. One can use them in creating their websites.